New data load (#381)
Things this PR does:

Adds cancel signal to all IQ methods. This ensures that when a query is no longer being observed, it will not attempt to fetch the data. This improvement prevents unnecessary data fetching and addresses an issue where fast scrolling would trigger data requests for elements that are no longer visible on the screen. This problem was present in the previous implementation but became more noticeable with the ability to make requests even for just one FFT worth of data in the case of scrolling.

Introduces the capability to download data in blocks and indexes. Users can specify the size of the IQ data block they want and the specific blocks they want to download.

Implements a hook that calculates the required data to be downloaded based on the current spectrogram parameters, taking into account the data that has already been downloaded.

Adds cache functionality to the hook, ensuring that only tiles that have not been seen before are downloaded. This optimization further improves data loading efficiency.